refactor(setup): isolate host launcher migration
Move Codex, Kiro, Copilot, and OpenCode host-launcher reconciliation into a dedicated setup_host module. Keep runtime-injected setup dependencies for transactional rollback while restoring setup.py and adapter module-size ratchets.\n\nAffected files:\n- src/ouroboros/cli/commands/setup.py\n- src/ouroboros/cli/commands/setup_host.py\n- src/ouroboros/mcp/server/adapter.py\n- tests/unit/cli/test_setup.py
