+++ src/ouroboros/cli/commands/setup.py
@@ -3909,298 +3909,13 @@ def _setup_runtime_only_backend(
     print_info(f"Config saved to: {config_path}")
 
 
-def _codex_mcp_entry_runtime_override(entry: dict[str, object]) -> str | None:
-    """Return an explicit non-host runtime selector from a Codex MCP entry."""
-    args = entry.get("args")
-    if isinstance(args, list):
-        for index, arg in enumerate(args):
-            if not isinstance(arg, str):
-                continue
-            if arg == "--runtime" and index + 1 < len(args):
-                value = args[index + 1]
-                if isinstance(value, str) and value != "host":
-                    return f"--runtime {value}"
-            elif arg.startswith("--runtime="):
-                value = arg.partition("=")[2]
-                if value != "host":
-                    return arg
-
-    env = entry.get("env")
-    if isinstance(env, dict):
-        for key in ("OUROBOROS_AGENT_RUNTIME", "OUROBOROS_RUNTIME"):
-            value = env.get(key)
-            if isinstance(value, str) and value.strip() and value.strip().lower() != "host":
-                return f"{key}={value}"
-    return None
-
-
-def _render_codex_host_mcp_section(entry: dict[str, object]) -> str:
-    """Render a setup-managed Codex entry after selecting host dispatch."""
-    migrated = deepcopy(entry)
-    args = migrated.get("args")
-    assert isinstance(args, list)
-    migrated_args = [str(arg) for arg in args]
-    for index, arg in enumerate(migrated_args):
-        if arg == "--runtime" and index + 1 < len(migrated_args):
-            migrated_args[index + 1] = "host"
-        elif arg.startswith("--runtime="):
-            migrated_args[index] = "--runtime=host"
-    migrated["args"] = migrated_args
-
-    env = migrated.get("env")
-    migrated_env = dict(env) if isinstance(env, dict) else {}
-    migrated_env["OUROBOROS_AGENT_RUNTIME"] = "host"
-    migrated["env"] = migrated_env
-
-    lines = [*_CODEX_MCP_COMMENT_LINES, "", "[mcp_servers.ouroboros]"]
-    for key in ("command", "args"):
-        lines.append(f"{key} = {_render_toml_value(migrated[key])}")
-    lines.extend(("", "[mcp_servers.ouroboros.env]"))
-    lines.extend(
-        f"{_render_toml_key(str(key))} = {_render_toml_value(value)}"
-        for key, value in migrated_env.items()
-    )
-    return "\n".join(lines) + "\n"
-
-
-def _migrate_codex_mcp_entry_to_host() -> bool:
-    """Migrate setup-owned Codex launch selectors, rejecting user-owned conflicts."""
-    codex_config = resolve_codex_home() / "config.toml"
-    if not codex_config.exists():
-        return True
-    try:
-        raw = codex_config.read_text(encoding="utf-8")
-        parsed = tomllib.loads(raw)
-    except (OSError, UnicodeDecodeError, tomllib.TOMLDecodeError) as exc:
-        print_error(f"Could not inspect {codex_config} before host activation: {exc}")
-        return False
-
-    entry = _codex_mcp_entry_from_toml(parsed)
-    if entry is None:
-        return True
-    has_managed_comment = _has_managed_codex_mcp_comment(raw)
-    if not _is_setup_managed_codex_mcp_entry(entry, has_managed_comment=has_managed_comment):
-        override = _codex_mcp_entry_runtime_override(entry)
-        if override is not None:
-            print_error(
-                f"Host setup cannot replace user-managed {codex_config}: its Ouroboros "
-                f"MCP entry explicitly selects {override}. Remove that selector or change it "
-                "to host, then rerun setup."
-            )
-            return False
-        return True
-
-    migrated, _ = _upsert_codex_mcp_section(raw, _render_codex_host_mcp_section(entry))
-    if migrated != raw:
-        _atomic_write_text(codex_config, migrated)
-        print_success(f"Migrated setup-managed Ouroboros MCP runtime to host in {codex_config}")
-    return True
-
-
-def _runtime_selector_override(
-    entry: dict[str, object],
-    *,
-    command_key: str,
-    env_key: str,
-    args_key: str | None = None,
-) -> str | None:
-    """Return the first explicit non-host runtime selector in a JSON MCP entry."""
-    command = entry.get(args_key) if args_key is not None else entry.get(command_key)
-    if isinstance(command, list):
-        for index, arg in enumerate(command):
-            if not isinstance(arg, str):
-                continue
-            if arg == "--runtime" and index + 1 < len(command):
-                value = command[index + 1]
-                if isinstance(value, str) and value.strip().lower() != "host":
-                    return f"--runtime {value}"
-            elif arg.startswith("--runtime=") and arg.partition("=")[2].strip().lower() != "host":
-                return arg
-
-    env = entry.get(env_key)
-    if isinstance(env, dict):
-        for key in ("OUROBOROS_AGENT_RUNTIME", "OUROBOROS_RUNTIME"):
-            value = env.get(key)
-            if isinstance(value, str) and value.strip() and value.strip().lower() != "host":
-                return f"{key}={value}"
-    return None
-
-
-def _migrate_json_mcp_entry_to_host(
-    *,
-    config_path: Path,
-    host_name: str,
-    servers_key: str,
-    command_key: str,
-    env_key: str,
-    setup_managed: Callable[[object], bool],
-    args_key: str | None = None,
-    load_jsonc: bool = False,
-) -> bool:
-    """Migrate one setup-owned JSON MCP launcher, rejecting user-owned conflicts."""
-    if not config_path.exists():
-        return True
-    try:
-        raw = config_path.read_text(encoding="utf-8")
-        data = json.loads(_strip_jsonc(raw) if load_jsonc else raw)
-    except (OSError, UnicodeDecodeError, json.JSONDecodeError) as exc:
-        print_error(f"Could not inspect {config_path} before host activation: {exc}")
-        return False
-    if not isinstance(data, dict):
-        print_error(
-            f"Could not inspect {config_path} before host activation: top-level is not an object"
-        )
-        return False
-    servers = data.get(servers_key)
-    if not isinstance(servers, dict):
-        return True
-    entry = servers.get("ouroboros")
-    if not isinstance(entry, dict):
-        return True
-
-    override = _runtime_selector_override(
-        entry,
-        command_key=command_key,
-        env_key=env_key,
-        args_key=args_key,
-    )
-    if not setup_managed(entry.get(command_key)):
-        if override is not None:
-            print_error(
-                f"Host setup cannot replace user-managed {config_path}: its Ouroboros "
-                f"MCP entry explicitly selects {override}. Remove that selector or change it "
-                "to host, then rerun setup."
-            )
-            return False
-        return True
-
-    migrated = deepcopy(entry)
-    runtime_args_key = args_key or command_key
-    command = migrated.get(runtime_args_key)
-    if isinstance(command, list):
-        migrated_command = [str(arg) for arg in command]
-        for index, arg in enumerate(migrated_command):
-            if arg == "--runtime" and index + 1 < len(migrated_command):
-                migrated_command[index + 1] = "host"
-            elif arg.startswith("--runtime="):
-                migrated_command[index] = "--runtime=host"
-        migrated[runtime_args_key] = migrated_command
-
-    env = migrated.get(env_key)
-    migrated_env = dict(env) if isinstance(env, dict) else {}
-    for key in ("OUROBOROS_AGENT_RUNTIME", "OUROBOROS_RUNTIME"):
-        if key in migrated_env:
-            migrated_env[key] = "host"
-    migrated_env["OUROBOROS_AGENT_RUNTIME"] = "host"
-    migrated[env_key] = migrated_env
-    if migrated == entry:
-        return True
-
-    servers["ouroboros"] = migrated
-    _atomic_write_text(config_path, json.dumps(data, indent=2) + "\n")
-    print_success(f"Migrated setup-managed {host_name} MCP runtime to host in {config_path}")
-    return True
-
-
-def _migrate_setup_managed_mcp_launchers_to_host() -> bool:
-    """Reconcile every setup-managed launcher whose env can override host config."""
-    known_commands = {"uvx", "pipx", "ouroboros", "python3", "python", "uv"}
-
-    def setup_managed_string(command: object) -> bool:
-        return isinstance(command, str) and (
-            command in known_commands
-            or os.path.basename(command) in {"ouroboros", "python3", "python"}
-        )
-
-    def setup_managed_array(command: object) -> bool:
-        return (
-            isinstance(command, list)
-            and bool(command)
-            and isinstance(command[0], str)
-            and (
-                command[0] in known_commands
-                or os.path.basename(command[0]) in {"ouroboros", "python3", "python"}
-            )
-        )
-
-    migrations = (
-        _migrate_codex_mcp_entry_to_host,
-        lambda: _migrate_json_mcp_entry_to_host(
-            config_path=Path.home() / ".kiro" / "settings" / "mcp.json",
-            host_name="Kiro",
-            servers_key="mcpServers",
-            command_key="command",
-            env_key="env",
-            setup_managed=setup_managed_string,
-            args_key="args",
-        ),
-        lambda: _migrate_json_mcp_entry_to_host(
-            config_path=Path.home() / ".copilot" / "mcp-config.json",
-            host_name="Copilot",
-            servers_key="mcpServers",
-            command_key="command",
-            env_key="env",
-            setup_managed=setup_managed_string,
-            args_key="args",
-        ),
-        lambda: _migrate_json_mcp_entry_to_host(
-            config_path=_find_opencode_config(),
-            host_name="OpenCode",
-            servers_key="mcp",
-            command_key="command",
-            env_key="environment",
-            setup_managed=setup_managed_array,
-            load_jsonc=True,
-        ),
-    )
-    return all(migrate() for migrate in migrations)
-
-
 def _setup_host() -> bool:
-    """Configure the CLI-less host runtime and its setup-owned MCP launcher."""
-    from ouroboros.config.loader import create_default_config, ensure_config_dir
-    from ouroboros.config.models import get_default_config
+    """Configure the CLI-less host runtime and its setup-owned MCP launchers."""
+    import sys
 
-    config_dir = ensure_config_dir()
-    config_path = config_dir / "config.yaml"
-    config_was_missing = not config_path.exists()
-    if config_was_missing:
-        config_dict = get_default_config().model_dump(mode="json")
-    else:
-        config_dict = yaml.safe_load(config_path.read_text(encoding="utf-8")) or {}
-    if not isinstance(config_dict, dict):
-        print_error("~/.ouroboros/config.yaml top-level is not a mapping — aborting host setup.")
-        return False
+    from ouroboros.cli.commands.setup_host import setup_host
 
-    orch = config_dict.get("orchestrator")
-    if not isinstance(orch, dict):
-        orch = {}
-        config_dict["orchestrator"] = orch
-    orch["runtime_backend"] = "host"
-
-    if not _commit_runtime_activation(
-        runtime_name="Host",
-        host_path=resolve_codex_home() / "config.toml",
-        additional_host_paths=(
-            Path.home() / ".kiro" / "settings" / "mcp.json",
-            Path.home() / ".copilot" / "mcp-config.json",
-            _find_opencode_config(),
-        ),
-        config_path=config_path,
-        config_was_missing=config_was_missing,
-        runtime_content=yaml.safe_dump(config_dict, default_flow_style=False, sort_keys=False),
-        register_host=_migrate_setup_managed_mcp_launchers_to_host,
-        create_defaults=create_default_config,
-    ):
-        return False
-
-    print_success("Configured host runtime (no CLI — host-driven dispatch)")
-    print_info(f"Config saved to: {config_path}")
-    print_info(
-        "'host' only works from an MCP host session pumping "
-        "ouroboros_job_wait (e.g. dsh). A terminal `ooo run` rejects it."
-    )
-    return True
+    return setup_host(sys.modules[__name__])
 
 
 def _setup_antigravity(antigravity_path: str) -> None:
